+++ src/pages/chart/scripts/staticData/index.ts
