+++ src/pages/chart/stores/chartStore.ts
@@ -5,6 +5,7 @@ import {
 } from 'src/scripts/databaseRepositories'
 import { Refuel, Vehicle } from 'src/scripts/libraries/refuel/models'
 import { ref } from 'vue'
+import { DataSource, Group } from '../models/models'
 
 export const useChartStore = defineStore('chartStore', () => {
   const vehicle = ref<Vehicle | null>(null)
@@ -23,6 +24,11 @@ export const useChartStore = defineStore('chartStore', () => {
       untilDate.value
     )
   }
+
+  function getChartData(dataSource: DataSource, groupBy: Group) {
+    return []
+  }
+
   return {
     vehicle,
     refuels,
@@ -30,6 +36,7 @@ export const useChartStore = defineStore('chartStore', () => {
     untilDate,
     groupBy,
     dataSource,
-    readData
+    readData,
+    getChartData
   }
 })
